Finalize component and delete abstract class
